Split `tool.rs` into modules
